Lab01.2/ddm501-lab2-starter/requirements.txt
# Core ML libraries
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
scikit-surprise==1.1.3

# MLflow for experiment tracking
mlflow==2.9.2

# Airflow for orchestration
apache-airflow==2.8.0

# FastAPI (from Lab 1)
fastapi==0.104.1
uvicorn==0.24.0

# Data validation
pydantic==2.5.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1
click==8.1.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Database
sqlalchemy==2.0.23
